Add support for scanning to launch from background
Errors should be short, but file locations are long

Perhaps one of the most misleading errors yet...

Routine cleanup and address Studio warnings

+++ app/src/main/java/com/hiddenramblings/tagmo/WebActivity.java
@@ -31,8 +31,8 @@ import androidx.webkit.WebViewAssetLoader.AssetsPathHandler;
 import androidx.webkit.WebViewClientCompat;
 import androidx.webkit.WebViewFeature;
 
-import com.eightbit.io.Debug;
-import com.eightbit.os.Storage;
+import com.hiddenramblings.tagmo.eightbit.io.Debug;
+import com.hiddenramblings.tagmo.eightbit.os.Storage;
 import com.hiddenramblings.tagmo.TagMo.Website;
 import com.hiddenramblings.tagmo.amiibo.AmiiboManager;
 import com.hiddenramblings.tagmo.nfctech.TagUtils;
@@ -173,7 +173,7 @@ public class WebActivity extends AppCompatActivity {
                         + "; charset=" + TagMo.UTF_8.displayName(), null);
             } catch (Exception e) {
                 Debug.Log(e);
-                new Toasty(this).Long(R.string.fail_logcat);
+                new Toasty(this).Short(R.string.fail_logcat);
                 finish();
             }
         }
